feat(files): add public subpath and wire JS to native renames
