+++ src/index.tsx
@@ -19,7 +19,7 @@ export {
   resolveModelPath,
 } from './utils';
 
-export { copyFileToContentUri } from './tts';
+export { copyFileToContentUri } from './files';
 export * from './alignment';
 
 export { getModelLicenses, type ModelLicense } from './licenses';
@@ -28,6 +28,7 @@ export { getModelLicenses, type ModelLicense } from './licenses';
 // - import { createTTS, ... } from 'react-native-sherpa-onnx/tts'
 // - import { detectAlignmentModel, ... } from 'react-native-sherpa-onnx/alignment'
 // - import { ... } from 'react-native-sherpa-onnx/download'
+// - import { copyFileToContentUri, ... } from 'react-native-sherpa-onnx/files'
 // - import { getBundledArchives, listBundledArchives, extractArchive } from 'react-native-sherpa-onnx/extraction'
 // - import { ... } from 'react-native-sherpa-onnx/vad'
 // - import { ... } from 'react-native-sherpa-onnx/diarization'
